[critical] Removed "all" outdated classes and packages in unit selection: Package de.dfki.lt.mary.unitselection.clunits Package de.dfki.lt.mary.unitselection.voiceimport Classes PathExtractor, PathExtractorImpl
Minor cleanups in other code to remove remaining compile errors.

git-svn-id: https://mary.opendfki.de/repos/trunk@295 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/viterbi/ViterbiCandidate.java
@@ -44,14 +44,14 @@ public class ViterbiCandidate {
     private ViterbiPath bestPath = null;
 	private Target target = null;
 	private ViterbiCandidate next = null;
-	private int targetCost = -1;
+	private double targetCost = -1;
 	
 	/**
 	 * Calculates and returns the target cost for this candidate
 	 * @param tcf the target cost function 
 	 * @return the target cost
 	 */
-	public int getTargetCost(TargetCostFunction tcf){
+	public double getTargetCost(TargetCostFunction tcf){
 	    if (targetCost == -1){
 	        targetCost = tcf.cost(target,unit);
 	        //System.out.println("Candidate with target "+target.getName()
